feat: increase time entries loading capabilities ✨
