+++ server/graphql/resolvers/index.ts
@@ -6,7 +6,7 @@ import { NotificationResolver } from './notification'
 import { OutlookCategoryResolver } from './outlookCategory'
 import { ProjectResolver } from './project/ProjectResolver'
 import { ReportLinkResolver } from './reportLink'
-import { ReportsResolver } from './reports'
+import { ReportsResolver } from './reports/ReportsResolver'
 import { RoleResolver } from './role'
 import { SubscriptionResolver } from './subscription'
 import { TimesheetResolver } from './timesheet'
@@ -33,7 +33,7 @@ export { NotificationResolver } from './notification'
 export { OutlookCategoryResolver } from './outlookCategory'
 export { ProjectResolver } from './project'
 export { ReportLinkResolver } from './reportLink'
-export { ReportsResolver } from './reports'
+export { ReportsResolver } from './reports/ReportsResolver'
 export { RoleResolver } from './role'
 export { SubscriptionResolver } from './subscription'
 export { TimesheetResolver } from './timesheet'
